Move substitute to a leaf module to break the helpers import cycle
`helpers/string` imported `deprecatedWarnOnce` from `helpers/console`, which
already imported `substitute` from `helpers/string`. It worked only because
both uses sit inside function bodies, but `helpers/console` has two
module-level initializers, so the first top-level call to `substitute` there
would leave one module reading an uninitialized binding and throw at import
time rather than at call time. `import/no-cycle` is off, so nothing flagged it.

`substitute` now lives in `helpers/templateString`, a leaf module both can
import. `helpers/string` re-exports it, so its documented public path and the
`typeof import('./helpers/string')` helper typing are unchanged.
